back_end_material_manager/src/requirements.txt
asgiref==3.6.0
boto3==1.26.156
botocore==1.29.156
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.6
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
#mysql-client==0.0.1
#mysql-connector-python==8.0.33
#mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==23.0
protobuf==3.20.3
#psycopg2==2.9.6
psycopg2-binary==2.9.6
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
#ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.5.0
